requirements.txt
kornia==0.7.2
pytorch-lightning==2.1
torchdata==0.7.1
invisible-watermark==0.2.0
fairscale>=0.4.13
open-clip-torch>=2.20.0
clip@git+https://github.com/openai/CLIP.git
timm>=0.9.2
torchmetrics>=1.0.1
black==23.7.0
chardet==5.1.0
einops>=0.6.1
fire>=0.5.0
fsspec>=2023.6.0
matplotlib>=3.7.2
natsort>=8.4.0
ninja>=1.11.1
numpy>=1.24.4
omegaconf>=2.3.0
opencv-python==4.6.0.66
pandas>=2.0.3
pillow>=9.5.0
pudb>=2022.1.3
pyyaml>=6.0.1
rembg
scipy>=1.10.1
streamlit>=0.73.1
tensorboardx==2.6
tokenizers==0.12.1
tqdm>=4.65.0
transformers==4.19.1
urllib3<1.27,>=1.25.4
wandb>=0.15.6
webdataset>=0.2.33
wheel>=0.41.0
gradio>=5.9.0
streamlit-keyup==0.2.0
nuscenes-devkit
